fix: Correções na edição da categoria

+++ src/Controllers/ListingCategoryController.php
@@ -118,7 +118,11 @@ class ListingCategoryController {
 
             if(!empty($iconData)) {
                 $iconName = $category->getId() . "_icon." . $iconData['mimeType'];
-                UploadFile::removeFile($category->getIcon(), $this->categoryImagesFolder);
+
+                if($category->getIcon()) {
+                    UploadFile::removeFile($category->getIcon(), $this->categoryImagesFolder);
+                }
+
                 UploadFile::uploadFile($files['icon'], $this->categoryImagesFolder, $iconName);
                 $category->setIcon($iconName);
             }
@@ -130,62 +134,6 @@ class ListingCategoryController {
                 "categoryData" => $category->toArray()
             ], 201);
 
-        } catch (Exception $e) {
-            logError($e->getMessage());
-            return $response->json([
-                'message' => FATAL_ERROR
-            ], 500);
-        }
-    }
-
-    public function updateIcon(Request $request, Response $response, $params) {
-        try {
-            $files = $request->files();
-            $iconData = "";
-            $photoData = "";
-
-            if(!empty($files['icon']) || !empty($files['photo'])) {
-                $errors = "";
-
-                if(!empty($files['icon'])) {
-                    $iconData = FileValidators::validIcon($files['icon']);
-
-                    if(isset($iconData['invalid'])) {
-                        $errors = $iconData['invalid'];
-                    }
-                }
-
-                if(!empty($errors)) {
-                    return $response->json([
-                        'message' => $errors
-                    ], 400);
-                }
-
-                $category = new ListingCategory();
-                $category->fetchById($params[0]);
-
-                if(!empty($iconData)) {
-                    $iconName = $category->getId() . "_icon." . $iconData['mimeType'];
-                    
-                    if(!empty($category->getIcon())) {
-                        UploadFile::removeFile($category->getIcon(), $this->categoryImagesFolder);
-                    }
-
-                    UploadFile::uploadFile($files['icon'], $this->categoryImagesFolder, $iconName);
-                    $category->setIcon($iconName);
-                }
-
-                $category->save();
-
-                return $response->json([
-                    "message" => CATEGORY_UPDATED,
-                ], 201);
-            } else {
-                return $response->json([
-                    "message" => NO_FILES,
-                ], 400);
-            }
-            
         } catch (Exception $e) {
             logError($e->getMessage());
             return $response->json([
